+++ app/renderer/types.js
@@ -1,32 +1,60 @@
-const apiActionTypes = name => {
-  const prefix = name.toUpperCase();
-  return {
-    requested: `${prefix}_REQUESTED`,
-    completed: `${prefix}_COMPLETED`,
-    failed: `${prefix}_FAILED`,
-    cancelled: `${prefix}_CANCELLED`,
-  };
-};
+import getActionTypeCreators from 'renderer/helpers/typeCreator';
 
-const changedActionTypes = name => {
-  const prefix = name.toUpperCase();
-  return {
-    changeRequested: `${prefix}_CHANGE_REQUESTED`,
-    changed: `${prefix}_CHANGED`,
-  };
-};
+const ACTION_TYPES_NAME_SPACE = 'ODIN';
 
-const triggerActionTypes = name => {
-  const prefix = name.toUpperCase();
-  return {
-    triggered: `${prefix}_TRIGGERED`,
-  };
-};
+const {
+  apiActionTypes,
+  changedActionTypes,
+  triggerActionTypes,
+  toggledActionTypes,
+  subscriptionActionTypes,
+} = getActionTypeCreators(ACTION_TYPES_NAME_SPACE);
 
 const actionTypes = {
   testPage: triggerActionTypes('test_page'),
-  updateMainNetBestBlock: triggerActionTypes('update_main_net_best_block'),
-  updateLocalNetBestBlock: triggerActionTypes('update_local_net_best_block'),
+  homePageLoad: triggerActionTypes('home_page_load'),
+  walletRestorePageLoad: triggerActionTypes('wallet_restore_page_load'),
+  navigation: triggerActionTypes('navigation'),
+  acceptTermsOfUse: triggerActionTypes('accept_terms_of_use'),
+  resetTermsOfUse: triggerActionTypes('reset_terms_of_use'),
+  storeSelectedNetwork: triggerActionTypes('store_selected_network'),
+
+  /* node system */
+  nodeJsonRpcSystem: apiActionTypes('node_jsonrpc_system'),
+  nodeJsonRpcSystemVersion: apiActionTypes('node_jsonrpc_system_version'),
+  nodeJsonRpcSystemChain: apiActionTypes('node_jsonrpc_system_chain'),
+  nodeJsonRpcSystemName: apiActionTypes('node_jsonrpc_system_name'),
+  nodeJsonRpcSystemHealth: apiActionTypes('node_jsonrpc_system_health'),
+
+  nodeWsChainSubscribeNewHead: apiActionTypes('node_ws_chain_subscribeNewHead'),
+
+  /* Stream */
+  stream: apiActionTypes('stream'),
+  streamStatus: changedActionTypes('stream_status'),
+  streamPing: apiActionTypes('stream_ping'),
+  streamMessage: changedActionTypes('stream_message'),
+  streamError: changedActionTypes('stream_error'),
+
+  /* Remote Stream */
+  remoteStream: apiActionTypes('remote_stream'),
+  remoteStreamStatus: changedActionTypes('remote_stream_status'),
+  remoteStreamPing: apiActionTypes('remote_stream_ping'),
+  remoteStreamMessage: changedActionTypes('remote_stream_message'),
+  remoteStreamError: changedActionTypes('remote_stream_error'),
+
+  /* Sync Stream */
+  syncStream: apiActionTypes('sync_stream'),
+  syncStreamStatus: changedActionTypes('sync_stream_status'),
+  syncStreamPing: apiActionTypes('sync_stream_ping'),
+  syncStreamMessage: changedActionTypes('sync_stream_message'),
+  syncStreamError: changedActionTypes('sync_stream_error'),
+
+  /* Sync Remote Stream */
+  syncRemoteStream: apiActionTypes('sync_remote_stream'),
+  syncRemoteStreamStatus: changedActionTypes('sync_remote_stream_status'),
+  syncRemoteStreamPing: apiActionTypes('sync_remote_stream_ping'),
+  syncRemoteStreamMessage: changedActionTypes('sync_remote_stream_message'),
+  syncRemoteStreamError: changedActionTypes('sync_remote_stream_error'),
 };
 
 export default actionTypes;