+++ CHANGELOG.md
@@ -8,10 +8,26 @@
 
 ### Improvements
 
+## v0.2.2 (2019.01.13)
+
+### Features
+
+### Fixes
+
+- fix hot reload issue
+
+### Improvements
+
 - Able launch app with runtime parameter
 - Able to launch app with cennznet-node in windows
 - Add ./scripts/init.sh to init dev env
 - Add Select, Input, Spinner components into Styleguidist
+- able to make network request
+- able to restart cennznet node with options
+- Connect TopBarContainer to redux state 
+- Add JsonRPC support
+- Add Websocket support
+- Able to display network status
 
 ## v0.2.1 (2019.01.04)